Refactor out algorithm better
